Use 10.4 SDK, not 10.6, as Mac OS 10.6 does not support PowerPC
Without this, what's the point of building Universal binaries
in the first place?

+++ other/TransparentWindow.m
@@ -54,7 +54,7 @@
 @implementation TransparentWindow
 
 - (id)initWithContentRect:(NSRect)contentRect
-				styleMask:(NSUInteger)aStyle
+				styleMask:(unsigned int)aStyle
 				  backing:(NSBackingStoreType)bufferingType
 					defer:(BOOL)flag
 {
